Base template: switch navigation to use path hierarchy (closes #17)
Adds a path_hierarchy context processor which exposes a template 
variable `PATH_HIERARCHY` which is the request path split on `/`.
This allows the navbar to simply check the name of the first
component rather than requiring each view to provide the name of the
active section and discourages using different names in the URLs and
elsewhere internally.