Base template: add body_classes block
This provides an easy way to add extra classes to the <body> tag as
needed and is set initially with a section-[FIRST PATH COMPONENT]
value to make it particularly easy to root styles under a selector
which starts with e.g. .section-events. 

(At some point, this will be even cleaner if we start using
something like LESS which supports nesting)